generichelpers cleanup
